chinese_checkers/requirements.txt
absl-py==1.3.0
chex==0.1.5
cloudpickle==2.2.1
contourpy==1.0.6
cycler==0.11.0
dm-haiku @ git+https://github.com/deepmind/dm-haiku@9c9624a58a700a8f2827ff073fbe9e65badc8606
dm-tree==0.1.8
fonttools==4.38.0
install==1.3.5
jax==0.4.1
jaxlib==0.4.1
Jinja2==3.1.2
jmp==0.0.2
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.2
mdurl==0.1.2
numpy==1.23.0
opt-einsum==3.3.0
optax==0.1.4
packaging==22.0
Pillow==9.3.0
pybind11 @ file:///Users/runner/miniforge3/conda-bld/pybind11-split_1679012568311/work
pybind11-global @ file:///Users/runner/miniforge3/conda-bld/pybind11-split_1679012568311/work
Pygments==2.14.0
pynvml==11.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
rich==13.3.1
scalene==1.5.19
scipy==1.9.3
six==1.16.0
tabulate==0.9.0
toolz==0.12.0
typing_extensions==4.4.0
